refactor: tesseract logic is in custom hook
